// controls/charts/src/sankey/model/sankey-interface.ts
export interface SankeyEventArgs {

    /** Defines the event cancel status. */
    cancel: boolean;
}

/**
 * Represents event arguments for sankey chart export.
 *
 * @interface SankeyExportEventArgs
 */
export interface SankeyExportEventArgs extends SankeyEventArgs {

    /**
     * The width of the exported chart.
     */
    width: number;
    /**
     * The height of the exported chart.
     */
    height: number;
}

/**
 * Represents the arguments for the print event in a sankey chart.
 *
 * @interface SankeyPrintEventArgs
 */
export interface SankeyPrintEventArgs extends SankeyEventArgs {

    /** Specifies the HTML content to be printed. */
    htmlContent: Element;
}